requirements.txt
flask==2.0.1
pymongo==4.0.1
bcrypt==3.2.0
python-dotenv==0.19.1
flask-session==0.4.0
flask-socketio==5.3.0
eventlet==0.33.3